Payment with Stripe (#17)
* Implement Stripe payment integration and update invoice handling

* Add support for marking invoices as paid via Stripe webhook with fallback to invoice ID

* Enhance checkout session extraction by adding error handling for Stripe event deserialization

* Refactor price handling to use BigDecimal for improved precision in booking and invoice calculations

+++ src/main/java/com/example/eventplatform/controller/BookingController.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -42,7 +43,7 @@ public class BookingController {
             @RequestParam String planName,
             @RequestParam String organizerName,
             @RequestParam String eventDate,
-            @RequestParam Double price,
+            @RequestParam BigDecimal price,
             Model model) {
 
         Booking savedBooking = bookingService.confirmBooking(
